requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
boto3==1.34.0
reportlab==4.0.9
prometheus_client==0.20.0
requests==2.32.3
